requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
aiofiles
python-multipart
